move LintMiddleware out of contrib
