+++ docs/contrib/lint.rst
@@ -1,9 +1,7 @@
-==========================
 Lint Validation Middleware
 ==========================
 
-.. currentmodule:: werkzeug.contrib.lint
-
-.. automodule:: werkzeug.contrib.lint
-
-.. autoclass:: LintMiddleware
+.. warning::
+    ``werkzeug.contrib.lint`` has moved to
+    :mod:`werkzeug.middleware.lint`. The old import is deprecated as of
+    version 0.15 and will be removed in version 1.0.
