+++ docs/middleware/lint.rst
@@ -1,0 +1,1 @@
+.. automodule:: werkzeug.middleware.lint
